TFP-5064 Forbedringer etter review
